refactoring and add pagination

+++ sick/service/src/main/java/com/epam/esm/service/DoctorService.java
@@ -4,7 +4,7 @@ import com.epam.esm.dao.DoctorDao;
 import com.epam.esm.dao.IllnessDao;
 import com.epam.esm.entity.Doctor;
 import com.epam.esm.entity.Illness;
-import com.epam.esm.util.NumberGenerator;
+import com.epam.esm.exception.EntityIsNotExistException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,7 +23,11 @@ public class DoctorService {
     }
 
     public Doctor findById(Long id) {
-        return doctorDao.findById(id);
+        if (doctorDao.existsById(id)) {
+            return doctorDao.findById(id);
+        } else {
+            throw new EntityIsNotExistException("doctor is not exist");
+        }
     }
 
     public List<Doctor> findAll(List<String> filters, String sortBy, String order) {
@@ -32,20 +36,62 @@ public class DoctorService {
 
     @Transactional
     public void create(Doctor doctor) {
+        fillExistIllnesses(doctor);
         doctorDao.create(doctor);
     }
 
     @Transactional
     public void update(Doctor doctor) {
-        doctorDao.update(doctor);
+        if (doctorDao.existsById(doctor.getId())) {
+            fillExistIllnesses(doctor);
+            doctorDao.update(doctor);
+        } else {
+            throw new EntityIsNotExistException("doctor is not exist");
+        }
+    }
+
+    @Transactional
+    public void partialUpdate(Doctor doctor) {
+        if (doctorDao.existsById(doctor.getId())) {
+            Doctor doctorFromDB = doctorDao.findById(doctor.getId());
+            if (doctor.getFirstName() != null) {
+                doctorFromDB.setFirstName(doctor.getFirstName());
+            }
+            if (doctor.getLastName() != null) {
+                doctorFromDB.setLastName(doctor.getLastName());
+            }
+            if (doctor.getMiddleName() != null) {
+                doctorFromDB.setMiddleName(doctor.getMiddleName());
+            }
+            if (doctor.getPhoneNumber() != null) {
+                doctorFromDB.setPhoneNumber(doctor.getPhoneNumber());
+            }
+            if (doctor.getPricePerConsultation() != null) {
+                doctorFromDB.setPricePerConsultation(doctor.getPricePerConsultation());
+            }
+            if (doctor.getDateOfBirth() != null) {
+                doctorFromDB.setDateOfBirth(doctor.getDateOfBirth());
+            }
+            if (doctor.getIllnesses() != null) {
+                fillExistIllnesses(doctor);
+                doctorFromDB.setIllnesses(doctor.getIllnesses());
+            }
+            doctorDao.update(doctorFromDB);
+        } else {
+            throw new EntityIsNotExistException("doctor is not exist");
+        }
     }
 
     @Transactional
     public void delete(Long id) {
-        doctorDao.delete(id);
+        if (doctorDao.existsById(id)) {
+            doctorDao.delete(id);
+        } else {
+            throw new EntityIsNotExistException("doctor is not exist");
+        }
     }
 
-    public void fillExistIllnesses(Doctor doctor) {
+    private void fillExistIllnesses(Doctor doctor) {
         if (doctor.getIllnesses() != null) {
             doctor.getIllnesses().forEach(illness -> {
                 if (illnessDao.existsByName(illness.getName())) {
@@ -56,10 +102,6 @@ public class DoctorService {
                 }
             });
         }
-    }
-
-    public boolean isDoctorExist(Long id) {
-        return doctorDao.existsById(id);
     }
 
     public Doctor findByIdentificationNumber(String identificationNumber) {
